# docs/openapi-api-service.yaml
openapi: 3.0.3
info:
  title: Firewall Defense API Service
  description: |
    REST API for Firewall Defense game. Handles player authentication, profiles, 
    match completion, leaderboards, global health, MOTD, and game configuration.
    
    All endpoints under `/api/admin/*` require the `X-Admin-API-Key` header.
    Player endpoints require JWT authentication via `Authorization: Bearer <token>` header.
  version: 1.0.0
  contact:
    name: Into Games Workshop
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: ${API_SERVICE_URL}
    description: Production server (set via environment variable)
...
paths:
...
  /api/admin/players:
    get:
      tags:
        - Admin
      summary: Get paginated list of players
      description: Returns a paginated list of all players with their profile information
      security:
        - adminApiKey: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of players per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminPlayerEntry'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/admin/players/{uuid}:
    post:
      tags:
        - Admin
      summary: Update player profile (admin)
      description: Update a player's corporation name and/or tagline (admin moderation)
      security:
        - adminApiKey: []
      parameters:
        - name: uuid
          in: path
          required: true
          description: Player UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  description: Corporation name (optional, validated against game config)
                  nullable: true
                tagline:
                  type: string
                  description: Tagline (optional, validated against game config)
              required: []
      responses:
        '200':
          description: Player updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPlayerEntry'
        '400':
          description: Validation error
          content:
            text/plain:
              schema:
                type: string
                example: "Corporation name must be at least 1 character(s)"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Player not found
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/leaderboard:
    get:
      tags:
        - Leaderboard
      summary: Get leaderboard and flush information
      description: |
        Returns the top 3 players by cumulative total score along with leaderboard flush information
        (when it was last flushed and the flush interval).
      responses:
        '200':
          description: Leaderboard and flush info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
        '500':
          description: Failed to retrieve leaderboard

  # === FIREWALL ENDPOINTS ===
...
  /api/admin/game-config:
    post:
      tags:
        - Admin
      summary: Update game configuration
      security:
        - adminApiKey: []
      description: |
        Updates the game configuration for all difficulty levels.
        All difficulty levels (easy, medium, hard) must be provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameConfig'
      responses:
        '200':
          description: Game config updated successfully
        '400':
          description: Invalid config format or values
  /api/admin/leaderboard:
    get:
      tags:
        - Admin
      summary: Get full leaderboard
      security:
        - adminApiKey: []
      description: Returns all leaderboard entries (not just top 3)
      responses:
        '200':
          description: Full leaderboard retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminLeaderboardEntry'
        '500':
          description: Failed to retrieve leaderboard

    delete:
      tags:
        - Admin
      summary: Flush leaderboard
      security:
        - adminApiKey: []
      description: |
        Resets all leaderboard entries and player scores.
        Broadcasts an empty leaderboard update to all connected clients.
      responses:
        '200':
          description: Leaderboard flushed successfully
        '500':
          description: Failed to flush leaderboard
  /api/admin/leaderboard-flush-interval:
    get:
      tags:
        - Admin
      summary: Get leaderboard flush interval
      security:
        - adminApiKey: []
      responses:
        '200':
          description: Flush interval retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  flushIntervalMinutes:
                    type: integer
                    description: Interval in minutes between automatic flushes
                    example: 60

    post:
      tags:
        - Admin
      summary: Set leaderboard flush interval
      security:
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - flushIntervalMinutes
              properties:
                flushIntervalMinutes:
                  type: integer
                  description: Interval in minutes (must be at least 1)
                  minimum: 1
                  example: 60
      responses:
        '200':
          description: Flush interval updated successfully
        '400':
          description: Invalid interval value
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/enroll
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Admin API key for admin endpoints
  schemas:
...
    GameConfig:
      type: object
      properties:
        validation:
          type: object
          description: Profile validation settings (min/max lengths for corporation name and tagline)
          properties:
            corporationNameMinLength:
              type: integer
              minimum: 1
              description: Minimum length for corporation name
            corporationNameMaxLength:
              type: integer
              minimum: 1
              description: Maximum length for corporation name
            taglineMinLength:
              type: integer
              minimum: 1
              description: Minimum length for tagline
            taglineMaxLength:
              type: integer
              minimum: 1
              description: Maximum length for tagline
        easy:
          $ref: '#/components/schemas/DifficultyConfig'
        medium:
          $ref: '#/components/schemas/DifficultyConfig'
        hard:
          $ref: '#/components/schemas/DifficultyConfig'
    DifficultyConfig:
      type: object
      properties:
        holeCount:
          type: integer
          description: Number of vulnerability holes on the canvas
          minimum: 0
          example: 1
        spawnRate:
          type: integer
          description: Bug spawn interval in milliseconds
          minimum: 1
          example: 750
        maxSpeed:
          type: number
          format: float
          description: Maximum bug movement speed
          minimum: 0.1
          example: 2.0
        penalty:
          type: integer
          description: Score deducted per bug reaching a hole
          minimum: 0
          example: 10
        defenseBonus:
          type: integer
          description: Score awarded per bug pushed off-screen
          minimum: 0
          example: 5
        gameTimeSeconds:
          type: integer
          description: Match duration in seconds
          minimum: 1
          example: 60
    LeaderboardResponse:
      type: object
      properties:
        leaderboard:
          type: array
          description: Top 3 players by cumulative total score
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
        lastFlush:
          type: integer
          format: int64
          nullable: true
          description: Unix timestamp (milliseconds) of last flush, or null if never flushed
          example: 1762138000000
        flushIntervalMinutes:
          type: integer
          description: Interval in minutes between automatic flushes
          example: 60
    LeaderboardEntry:
      type: object
      properties:
        productName:
          type: string
          description: Product/Company name (optional)
          example: "Acme Corp"
        tagline:
          type: string
          description: Player tagline/slogan
          example: "Defender-cbd5"
        score:
          type: integer
          description: Cumulative total score
          example: 1830
    AdminLeaderboardEntry:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: Player UUID
        productName:
          type: string
          description: Corporation name (optional)
          example: "Acme Corp"
        tagline:
          type: string
          description: Player tagline/slogan
        color:
          type: string
          description: Player color (hex format)
          example: "#FFFFFF"
        score:
          type: integer
          description: Cumulative total score
    AdminPlayerEntry:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: Player UUID
        productName:
          type: string
          description: Corporation name
        tagline:
          type: string
          description: Player tagline/slogan
        color:
          type: string
          description: Player color (hex format)
          example: "#FFFFFF"
        totalScore:
          type: integer
          description: Cumulative total score